+++ Trunk/src/scripts/src/InstanceScripts/BloodFurnace/Instance_BloodFurnace.cpp
@@ -1991,7 +1991,7 @@ public:
         spells[3].info = dbcSpell.LookupEntry(STEALTH);
 		spells[3].targettype = TARGET_SELF;
 		spells[3].instant = true;
-		spells[3].cooldown = -1;
+		spells[3].cooldown = (uint32)-1;
 		spells[3].perctrigger = 0.0f;
 		spells[3].attackstoptimer = 1000;
